fix: move ThresholdsSettings.tsx from shared to visualizations/internal
