+++ src/visualization/components/internal/ThresholdsSettings.tsx
@@ -1,4 +1,6 @@
-import React, {useState, FC} from 'react'
+// Libraries
+import React, {useMemo, useEffect, FunctionComponent} from 'react'
+import {cloneDeep} from 'lodash'
 
 // Components
 import ThresholdSetting from 'src/shared/components/ThresholdSetting'
@@ -13,6 +15,7 @@ import {
 } from '@influxdata/clockface'
 
 // Utils
+import {useOneWayReducer} from 'src/shared/utils/useOneWayReducer'
 import {
   sortThresholds,
   validateThresholds,
@@ -27,37 +30,113 @@ interface Props {
   onSetThresholds: (thresholds: Color[]) => void
 }
 
-const ThresholdsSettings: FC<Props> = ({thresholds, onSetThresholds}) => {
-  const [errors, setErrors] = useState({})
+interface State {
+  thresholds: Color[]
+  inputs: {[thresholdID: string]: string}
+  errors: {[thresholdID: string]: string}
+  isValid: boolean
+  isDirty: boolean
+}
 
-  const appendThreshold = () => {
-    const defaultThreshold = addThreshold(thresholds)
+type Action =
+  | {type: 'COLOR_CHANGED'; id: string; name: string; hex: string}
+  | {type: 'VALUE_CHANGED'; id: string; value: string}
+  | {type: 'VALUE_BLURRED'; id: string}
+  | {type: 'THRESHOLD_REMOVED'; id: string}
+  | {type: 'THRESHOLD_ADDED'}
 
-    onSetThresholds([...thresholds, defaultThreshold])
+const reducer = (state: State, action: Action): State => {
+  switch (action.type) {
+    case 'COLOR_CHANGED': {
+      const {id, name, hex} = action
+
+      const thresholds = state.thresholds.map(threshold =>
+        threshold.id === id ? {...threshold, name, hex} : threshold
+      )
+
+      return {...state, thresholds, isDirty: true}
+    }
+
+    case 'VALUE_CHANGED': {
+      const {id, value} = action
+
+      const inputs = {...state.inputs, [id]: value}
+
+      return {...state, inputs, isDirty: true, isValid: false}
+    }
+
+    case 'VALUE_BLURRED': {
+      const thresholds = state.thresholds.map(threshold =>
+        threshold.id === action.id
+          ? {...threshold, value: parseFloat(state.inputs[action.id])}
+          : threshold
+      )
+
+      const errors = validateThresholds(thresholds)
+
+      const isValid = Object.values(errors).length === 0
+
+      return {...state, thresholds, errors, isValid}
+    }
+
+    case 'THRESHOLD_ADDED': {
+      const newThreshold = addThreshold(state.thresholds)
+
+      const thresholds = sortThresholds([...state.thresholds, newThreshold])
+
+      const inputs = {
+        ...state.inputs,
+        [newThreshold.id]: String(newThreshold.value),
+      }
+
+      return {...state, thresholds, inputs, isDirty: true}
+    }
+
+    case 'THRESHOLD_REMOVED': {
+      const thresholds = state.thresholds.filter(
+        threshold => threshold.id !== action.id
+      )
+
+      return {...state, thresholds, isDirty: true}
+    }
+
+    default:
+      const unknownAction: never = action
+      const unknownActionType = (unknownAction as any).type
+
+      throw new Error(
+        `unhandled action of type "${unknownActionType}" in ThresholdsSettings`
+      )
   }
+}
 
-  const updateThreshold = (id: string, value: Partial<Color>) => {
-    onSetThresholds(
-      thresholds.map(threshold => {
-        if (threshold.id !== id) {
-          return threshold
-        }
+const ThresholdsSettings: FunctionComponent<Props> = ({
+  thresholds,
+  onSetThresholds,
+}) => {
+  const initialState: State = useMemo(
+    () => ({
+      thresholds: sortThresholds(
+        cloneDeep(thresholds.filter(({type}) => type !== 'scale'))
+      ),
+      inputs: thresholds.reduce(
+        (acc, {id, value}) => ({...acc, [id]: String(value)}),
+        {}
+      ),
+      errors: {},
+      isDirty: false,
+      isValid: true,
+    }),
+    [thresholds]
+  )
 
-        return {
-          ...threshold,
-          ...value,
-        }
-      })
-    )
-  }
+  const [state, dispatch] = useOneWayReducer(reducer, initialState)
 
-  const removeThreshold = (id: string) => {
-    onSetThresholds(thresholds.filter(threshold => threshold.id !== id))
-  }
-
-  const onBlur = () => {
-    setErrors(validateThresholds(thresholds))
-  }
+  useEffect(() => {
+    if (state.isDirty && state.isValid) {
+      onSetThresholds(state.thresholds)
+    }
+  }, [state, onSetThresholds])
 
   return (
     <FlexBox
@@ -70,25 +149,44 @@ const ThresholdsSettings: FC<Props> = ({thresholds, onSetThresholds}) => {
         shape={ButtonShape.StretchToFit}
         icon={IconFont.Plus}
         text="Add a Threshold"
-        onClick={() => appendThreshold()}
+        onClick={() => dispatch({type: 'THRESHOLD_ADDED'})}
       />
-      {sortThresholds(thresholds).map(threshold => {
+      {state.thresholds.map(threshold => {
         const onChangeValue = value =>
-          updateThreshold(threshold.id, {value: parseFloat(value)})
+          dispatch({
+            type: 'VALUE_CHANGED',
+            id: threshold.id,
+            value,
+          })
 
         const onChangeColor = (name, hex) =>
-          updateThreshold(threshold.id, {name, hex})
+          dispatch({
+            type: 'COLOR_CHANGED',
+            id: threshold.id,
+            name,
+            hex,
+          })
 
-        const onRemove = () => removeThreshold(threshold.id)
+        const onRemove = () =>
+          dispatch({
+            type: 'THRESHOLD_REMOVED',
+            id: threshold.id,
+          })
+
+        const onBlur = () =>
+          dispatch({
+            type: 'VALUE_BLURRED',
+            id: threshold.id,
+          })
 
         return (
           <ThresholdSetting
-            key={`threshold-${threshold.id}`}
-            id={`threshold-${threshold.id}`}
+            key={threshold.id}
+            id={threshold.id}
             name={threshold.name}
             type={threshold.type}
-            value={'' + threshold.value}
-            error={errors[threshold.id]}
+            value={state.inputs[threshold.id]}
+            error={state.errors[threshold.id]}
             onBlur={onBlur}
             onRemove={onRemove}
             onChangeValue={onChangeValue}
